Started implementation of data entry
TODO: implement entry of game map.
In CreateMap view: have to be a table of map pieces that can be clicked and then save which one was clicked.
In Map partial view: render an empty table that column's and row's number in equel to the map size. The user can click on each fields. It he/she clicked on a map piece before, load that map piece to tha clicked field and store it to the map.