Map display is partly implemented
At game creation the map display is partly implemented: it shows a table with as many cells and columns as the map, but that's all.
Need to change the database a bit, to read and display easily a map. The needed changes' plans are in the documentation.
TODO: change the database (the data are deleted) structure and show the map and save data into it.