Show the directions in the map at game creation
Now, implemented to show some test representation of directions on the map on Map.chtml.
But we got some bugs: ExitRoads weren't included to the Map, so when we would like to get it, got NullReferenceException. I don't know, who we could include it, but if we put those 4 boollean attributes to the field, we can get those directions. So I should do that. I thinks it will make the effectiveness better too.
TODO:
put ExitRoads' attributes into Field.
1.) Delete all data from database.
2.) Change the database formation, migrate it and ect.