WayDirections include bug solved and pictrues added
The bug solved: ExitRoads atributes added to Field class and ExitRoads class is removed. (The data table too.)
Some test map piece pictures added and loaded to database. (In SeedTestData.)
Now, we show these pictures to represent the way directions at map creation instead of some text.
Problem: the whole wesite is reloaded each time we click on a picture to fill the map. So, it's really slow. We should only reload that one picture which has been changed instead of the whole website.