Unavailable to play or edit if not logged in
Implemented to not show link for play if a user isn't logged in. (It wasn't available before too, but the program crashed if someone tryed: now, it's not crash, instead of it a message appears that inform the user that he/she must log in is want to play or edit games and the links for play or edit don't appears for him/her.